Break up Types
