+++ src/types/modifiers.ts
@@ -1,47 +1,13 @@
-/**
- * Options for configuring every option of a Randsum roll.
- *
- * @typeParam D Making overloading a breeze, always a Boolean.
- */
-export type RandsumOptions<D extends boolean = boolean> = RollOptions & UserOptions<D>
+import { NumberString } from './primitives'
 
 /**
- * Options for configuring the Dice Roll, without the sides parameter.
+ * An object to hold options for `Drop` modifiers.
  *
- * @typeParam D Making overloading a breeze, always a Boolean.
+ * {@link DropOptions}
  */
-export type RandsumOptionsWithoutSides<D extends boolean = boolean> = Omit<RandsumOptions<D>, 'sides'>
-
-/**
- * Options provided to the user not directly related to the dice roll:
- *
- * @typeParam D Making overloading a breeze, always a Boolean.
- */
-export interface UserOptions<D extends boolean = boolean> {
-  /** Whether or not to display a {@link RollResult} (true) or a `number` (false) */
-  detailed?: D
-  /** A custom functtion that replaces the default randomizer function */
-  randomizer?: (sides: NumberString) => number
+export interface DropModifier<T extends number | 'inclusive' = 'inclusive'> {
+  drop: DropOptions<T>
 }
-/**
- * `DiceNotation` is a [Template Literal](https://www.typescriptlang.org/docs/handbook/2/template-literal-types.html)
-patterned ater the basic dice notation (with room for extensions).
- * See [Randsum Dice Notation](https://github.com/alxjrvs/randsum/blob/main/RANDSUM_DICE_NOTATION.md) for more.
- */
-export type DiceNotation = `${number}${'d' | 'D'}${number}${string}`
-
-/**
- * NumberString leverages [Conditional Types](https://www.typescriptlang.org/docs/handbook/2/conditional-types.html) to simplify our interfaces.
- * When `Numberstring<'inclusive'>`, NumberString resolves to `number | ${number}` (a number or number-like string)
- *
- * When `NumberString<number>`, NumberString resolves to `number`.
- *
- * This allows us to use `RollOptions<number>` as the base for RollParameters, while
- still benefitting from strong types.
- */
-export type NumberString<T extends number | 'inclusive' = 'inclusive'> = T extends 'inclusive'
-  ? number | `${number}`
-  : number
 
 /**
  * Options for `Drop` modifiers.
@@ -62,14 +28,6 @@ export interface DropOptions<T extends number | 'inclusive' = 'inclusive'> {
 }
 
 /**
- * An object to hold options for `Drop` modifiers.
- *
- * {@link DropOptions}
- */
-export interface DropModifier<T extends number | 'inclusive' = 'inclusive'> {
-  drop: DropOptions<T>
-}
-
 /**
  * Options for `Cap` modifiers.
  *
@@ -178,7 +136,6 @@ export interface PlusModifier<T extends number | 'inclusive' = 'inclusive'> {
 export interface MinusModifier<T extends number | 'inclusive' = 'inclusive'> {
   minus: NumberString<T>
 }
-
 /**
  * All Possible modifiers used to modify the roll results
  *
@@ -192,44 +149,3 @@ export type Modifier<T extends number | 'inclusive' = 'inclusive'> =
   | UniqueModifier<T>
   | PlusModifier<T>
   | MinusModifier<T>
-
-/**
- * All Options available to be passed to {@link randsum}.
- *
- */
-export interface RollOptions<T extends number | 'inclusive' = 'inclusive'> {
-  quantity?: NumberString<T>
-  sides: NumberString<T>
-  modifiers?: Array<Modifier<T>>
-}
-
-/**
- * All parameters used to calculate a roll.
- *
- * Returned in {@link RollResult}
- *
- */
-
-export interface InternalRollParameters extends RollOptions<number>, UserOptions {
-  quantity: number
-  detailed: boolean
-  modifiers: Array<Modifier<number>>
-}
-
-export interface RollParameters extends InternalRollParameters {
-  initialRolls: number[]
-  rollOne: () => number
-}
-
-/**
- *
- * The Return value of a Detailed `randsum` roll.
- */
-export interface RollResult {
-  /** The sum total of the dice rolls, after modifiers have been applied */
-  total: number
-  /** An array of the numbers of each individual die rolled, after modifiers have been applied */
-  rolls: number[]
-  /** The object used the calculate the total and rolls */
-  rollParameters: RollParameters
-}
