+++ src/types/options.ts
@@ -1,0 +1,38 @@
+import { Modifier } from './modifiers'
+import { NumberString } from './primitives'
+
+/**
+ * Options for configuring every option of a Randsum roll.
+ *
+ * @typeParam D Making overloading a breeze, always a Boolean.
+ */
+export type RandsumOptions<D extends boolean = boolean> = RollOptions & UserOptions<D>
+
+/**
+ * Options for configuring the Dice Roll, without the sides parameter.
+ *
+ * @typeParam D Making overloading a breeze, always a Boolean.
+ */
+export type RandsumOptionsWithoutSides<D extends boolean = boolean> = Omit<RandsumOptions<D>, 'sides'>
+
+/**
+ * Options provided to the user not directly related to the dice roll:
+ *
+ * @typeParam D Making overloading a breeze, always a Boolean.
+ */
+export interface UserOptions<D extends boolean = boolean> {
+  /** Whether or not to display a {@link RollResult} (true) or a `number` (false) */
+  detailed?: D
+  /** A custom functtion that replaces the default randomizer function */
+  randomizer?: (sides: NumberString) => number
+}
+
+/**
+ * All Options available to be passed to {@link randsum}.
+ *
+ */
+export interface RollOptions<T extends number | 'inclusive' = 'inclusive'> {
+  quantity?: NumberString<T>
+  sides: NumberString<T>
+  modifiers?: Array<Modifier<T>>
+}
